+++ src/search_cmd_path.c
@@ -1,17 +1,17 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   search_cmd_path.c                                  :+:      :+:    :+:   */
+/*   new_search_cmd_path.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: yuotsubo <yuotsubo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/07/09 18:46:59 by yuotsubo          #+#    #+#             */
-/*   Updated: 2024/07/09 21:51:27 by yuotsubo         ###   ########.fr       */
+/*   Created: 2024/09/24 18:48:14 by yuotsubo          #+#    #+#             */
+/*   Updated: 2024/09/24 19:01:37 by yuotsubo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../pipex.h"
 #include "libft.h"
+#include "../pipex.h"
 
 static int	get_paths_num(char **paths)
 {
@@ -23,13 +23,14 @@ static int	get_paths_num(char **paths)
 	return (paths_num);
 }
 
-static t_bool	is_relative_path(char *cmd_name)
+static char *passed_path(char *cmd_name)
 {
-	if (!ft_strncmp(cmd_name, "./", ft_strlen("./")))
-		return (TRUE);
-	if (!ft_strncmp(cmd_name, "../", ft_strlen("../")))
-		return (TRUE);
-	return (FALSE);
+	char *res;
+
+	res = NULL;
+	if (!access(cmd_name, X_OK | F_OK))
+		res = ft_strdup(cmd_name);
+	return (res);
 }
 
 static char	**get_paths(char **environ)
@@ -100,22 +101,24 @@ static char	*get_cmd_path(char *cmd_name, char **paths)
 char	*search_cmd_path(char *cmd_name, char **environ)
 {
 	char	**paths;
-	char	*res;
+	char 	*res;
 
-	if (!access(cmd_name, X_OK | F_OK))
+	if (ft_strchr(cmd_name, '/'))
 	{
-		res = ft_strdup(cmd_name);
+		res = passed_path(cmd_name);
+		if (!res)
+			ft_printf("%s: No such file or directory\n", cmd_name);
 		return (res);
 	}
-	if (is_relative_path(cmd_name))
-		return (NULL);
 	paths = get_paths(environ);
-	if (!paths)
+	if(!paths)
 		return (NULL);
 	adjust_paths(paths);
 	if (!paths)
 		return (NULL);
 	res = get_cmd_path(cmd_name, paths);
+	if (!res)
+		ft_printf("%s: command not found\n", cmd_name);
 	free_paths(paths, get_paths_num(paths));
 	return (res);
 }
